Fixed stop condition bug; Extra interval; Readme
Stop condition wasn't to the spec of the Algo. - should stop only if the
number of intervals is equal to or greater than the number of classes
and the caim value is not improving.

Additional interval (interval[0] + (interval[1] - interval[0])/2.0) is
added to handle cases where best discretization requires an interval of
only the first value in the features sorted unique values.

Additional documentation added for interpretting the interval outputs.